requirements_gpu.txt
tensorflow-gpu
keras
pillow
scipy==1.2.1